Extract Event_Admin_List class from Event_Setup and Event_Query to consolidate admin list table logic.
