+++ includes/core/classes/class-event-query.php
@@ -15,7 +15,6 @@ namespace GatherPress\Core;
 // Exit if accessed directly.
 defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
 
-use GatherPress\Core\Rsvp;
 use GatherPress\Core\Settings;
 use GatherPress\Core\Traits\Singleton;
 use GatherPress\Core\Venue;
@@ -43,15 +42,6 @@ class Event_Query {
 	const EVENT_QUERY_PARAM = 'gatherpress_event_query';
 
 	/**
-	 * Cached event counts for the current request.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @var array<string, int>|null
-	 */
-	protected ?array $event_counts = null;
-
-	/**
 	 * Class constructor.
 	 *
 	 * This method initializes the object and sets up necessary hooks.
@@ -73,20 +63,8 @@ class Event_Query {
 	 */
 	protected function setup_hooks(): void {
 		add_action( 'pre_get_posts', array( $this, 'prepare_event_query_before_execution' ) );
-		add_action( 'pre_get_posts', array( $this, 'handle_rsvp_sorting' ) );
-		add_action( 'pre_get_posts', array( $this, 'handle_venue_sorting' ) );
 		// Priority 9 to run before the upcoming/past adjustments at priority 10.
 		add_filter( 'posts_clauses', array( $this, 'adjust_admin_event_sorting' ), 9, 2 );
-		add_action( 'load-edit.php', array( $this, 'default_sort' ) );
-		add_filter(
-			sprintf( 'manage_edit-%s_sortable_columns', Event::POST_TYPE ),
-			array( $this, 'sortable_columns' )
-		);
-		add_filter(
-			sprintf( 'views_edit-%s', Event::POST_TYPE ),
-			array( $this, 'views_edit' )
-		);
-		add_filter( 'query_vars', array( $this, 'query_vars' ) );
 	}
 
 	/**
@@ -539,414 +517,5 @@ class Event_Query {
 		// - Upcoming events, without running events.
 		// - Past events, that are still running.
 		return 'datetime_start_gmt';
-	}
-
-	/**
-	 * Sets the default sort field and sort order on the event post type admin screen, to order by event date.
-	 *
-	 * @author John Blackbourn @johnbillion
-	 * @source https://github.com/johnbillion/extended-cpts/blob/20b7e9773b60f7301cd59ee520affa0ff63f90e6/src/PostTypeAdmin.php#L160-L178
-	 *
-	 * @since 1.0.0
-	 *
-	 * @return void
-	 */
-	public function default_sort(): void {
-		$screen_id = sprintf( 'edit-%s', Event::POST_TYPE );
-		if ( ! function_exists( 'get_current_screen' ) || get_current_screen()->id !== $screen_id ) {
-			return;
-		}
-
-		// If the screen is already ordered, bail out.
-		if ( isset( $_GET['orderby'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
-			return;
-		}
-
-		// Default to sorting by event date ascending.
-		$_GET['orderby'] = 'datetime';
-		$_GET['order']   = 'asc';
-	}
-
-	/**
-	 * Make custom columns sortable for Event post type in the admin dashboard.
-	 *
-	 * This method allows the custom columns, including the 'Event date & time' and 'RSVPs' columns,
-	 * to be sortable in the WordPress admin dashboard for Event post types.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param array $columns An array of sortable columns.
-	 * @return array An updated array of sortable columns.
-	 */
-	public function sortable_columns( array $columns ): array {
-		// Add 'datetime' as a sortable column.
-		$columns['datetime'] = 'datetime';
-		// Add 'venue' as a sortable column.
-		$columns['venue'] = 'venue';
-		// Add 'rsvps' as a sortable column.
-		$columns['rsvps'] = 'rsvps';
-
-		return $columns;
-	}
-
-	/**
-	 * Add 'Upcoming' & 'Past' to the available admin event list table views.
-	 *
-	 * This method adds links to filter the shown events in the admin list,
-	 * the filtering allows to show 'upcoming' or 'past' events.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param array $view_links An array of available list table views.
-	 *
-	 * @return array Updated list table views.
-	 */
-	public function views_edit( array $view_links ): array {
-		// phpcs:ignore WordPress.Security.NonceVerification.Recommended
-		$current_view = isset( $_GET['gatherpress_event_query'] )
-			// phpcs:ignore WordPress.Security.NonceVerification.Recommended
-			? sanitize_text_field( wp_unslash( $_GET['gatherpress_event_query'] ) )
-			: '';
-
-		$counts    = $this->get_event_counts();
-		$placement = 1;
-		$inserts   = array(
-			'upcoming' => __( 'Upcoming', 'gatherpress' ),
-			'past'     => __( 'Past', 'gatherpress' ),
-		);
-		$base_url  = admin_url( 'edit.php' );
-
-		foreach ( $inserts as $key => $value ) {
-			$count           = isset( $counts[ $key ] ) ? $counts[ $key ] : 0;
-			$inserts[ $key ] = sprintf(
-				'<a href="%s"%s>%s <span class="count">(%s)</span></a>',
-				add_query_arg(
-					array(
-						'gatherpress_event_query' => $key,
-						'post_type'               => Event::POST_TYPE,
-						'order'                   => 'upcoming' === $key ? 'asc' : 'desc',
-						'orderby'                 => 'datetime',
-					),
-					$base_url
-				),
-				$key === $current_view ? ' class="current" aria-current="page"' : '',
-				$value,
-				number_format_i18n( $count )
-			);
-		}
-
-		if ( isset( $view_links['all'] ) ) {
-			if ( $current_view ) {
-				// Remove the "current" class from "All" when an event query filter is active.
-				$view_links['all'] = str_replace(
-					array( ' class="current"', ' aria-current="page"' ),
-					'',
-					$view_links['all']
-				);
-			} elseif ( false === strpos( $view_links['all'], 'class="current"' ) ) {
-				// Add "current" class to "All" when no filter is active.
-				// default_sort() adds orderby/order to $_GET which prevents
-				// WordPress from detecting this as a base request.
-				$view_links['all'] = str_replace(
-					'<a ',
-					'<a class="current" aria-current="page" ',
-					$view_links['all']
-				);
-			}
-		}
-
-		return array_slice( $view_links, 0, $placement, true )
-			+ $inserts
-			+ array_slice( $view_links, $placement, null, true );
-	}
-
-	/**
-	 * Get counts of upcoming and past events.
-	 *
-	 * Uses the same datetime comparison logic as Event_Query::adjust_event_sql()
-	 * with inclusive=true: upcoming uses datetime_end_gmt (includes running events),
-	 * past uses datetime_start_gmt (excludes running events).
-	 *
-	 * @since 1.0.0
-	 *
-	 * @return array<string, int> Associative array with 'upcoming' and 'past' counts.
-	 */
-	protected function get_event_counts(): array {
-		if ( null !== $this->event_counts ) {
-			return $this->event_counts;
-		}
-
-		global $wpdb;
-
-		$table   = sprintf( Event::TABLE_FORMAT, $wpdb->prefix );
-		$current = gmdate( Event::DATETIME_FORMAT, time() );
-
-		// Upcoming: events whose end time is still in the future (includes currently running),
-		// or events with no row in the events table (no date set yet).
-		// phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching
-		$upcoming = (int) $wpdb->get_var(
-			$wpdb->prepare(
-				'SELECT COUNT(1) FROM %i LEFT JOIN %i ON %i.ID = %i.post_id'
-				. ' WHERE %i.post_type = %s AND %i.post_status NOT IN'
-				. " ('trash', 'auto-draft') AND (%i.datetime_end_gmt >= %s"
-				. ' OR %i.post_id IS NULL)',
-				$wpdb->posts,
-				$table,
-				$wpdb->posts,
-				$table,
-				$wpdb->posts,
-				Event::POST_TYPE,
-				$wpdb->posts,
-				$table,
-				$current,
-				$table
-			)
-		);
-
-		// Past: events whose start time is in the past (excludes currently running),
-		// excluding events with no row in the events table.
-		// phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching
-		$past = (int) $wpdb->get_var(
-			$wpdb->prepare(
-				'SELECT COUNT(1) FROM %i LEFT JOIN %i ON %i.ID = %i.post_id'
-				. ' WHERE %i.post_type = %s AND %i.post_status NOT IN'
-				. " ('trash', 'auto-draft') AND %i.datetime_start_gmt < %s"
-				. ' AND %i.post_id IS NOT NULL',
-				$wpdb->posts,
-				$table,
-				$wpdb->posts,
-				$table,
-				$wpdb->posts,
-				Event::POST_TYPE,
-				$wpdb->posts,
-				$table,
-				$current,
-				$table
-			)
-		);
-
-		$this->event_counts = array(
-			'upcoming' => $upcoming,
-			'past'     => $past,
-		);
-
-		return $this->event_counts;
-	}
-
-	/**
-	 * Allowlist for additional query parameters.
-	 *
-	 * Adds 'gatherpress_event_query' to the list of allowed query variables,
-	 * to be able to request 'upcoming' or 'past' events in the admin list view.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param  string[] $query_vars List of allowed query variables.
-	 *
-	 * @return string[] Updated list of allowed query variables.
-	 */
-	public function query_vars( array $query_vars ) {
-		$query_vars[] = 'gatherpress_event_query';
-		return $query_vars;
-	}
-
-	/**
-	 * Handle RSVP column sorting in the events list.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param WP_Query $query The WP_Query instance.
-	 * @return void
-	 */
-	public function handle_rsvp_sorting( $query ): void {
-		$this->handle_column_sorting(
-			$query,
-			'rsvps',
-			array(
-				'posts_join_paged' => array( $this, 'rsvp_sorting_join_paged' ),
-				'posts_groupby'    => array( $this, 'sorting_groupby_post_id' ),
-				'posts_orderby'    => array( $this, 'rsvp_sorting_orderby' ),
-			),
-			'rsvp_sort_order'
-		);
-	}
-
-	/**
-	 * Handle venue sorting in the admin list table.
-	 *
-	 * This method modifies the query to sort events by venue name alphabetically.
-	 * Similar to how WordPress core handles taxonomy sorting.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param WP_Query $query The WP_Query instance.
-	 * @return void
-	 */
-	public function handle_venue_sorting( $query ): void {
-		$this->handle_column_sorting(
-			$query,
-			'venue',
-			array(
-				'posts_join_paged' => array( $this, 'venue_sorting_join_paged' ),
-				'posts_groupby'    => array( $this, 'sorting_groupby_post_id' ),
-				'posts_orderby'    => array( $this, 'venue_sorting_orderby' ),
-			),
-			'venue_sort_order'
-		);
-	}
-
-	/**
-	 * Handle column sorting for the admin events list table.
-	 *
-	 * Shared logic for sorting by custom columns (RSVP count, venue name, etc.).
-	 * Validates the query context, normalizes the sort order, registers the
-	 * provided SQL filter callbacks, and stores the order on the query object.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param WP_Query               $query     The WP_Query instance.
-	 * @param string                 $column    The column name to match against the orderby query var.
-	 * @param array<string,callable> $filters   Map of WordPress filter hook => callback to register.
-	 * @param string                 $order_key The query var key used to store the validated sort order.
-	 * @return void
-	 */
-	private function handle_column_sorting( WP_Query $query, string $column, array $filters, string $order_key ): void {
-		// Only proceed if we're in admin, on the main query, and dealing with events.
-		if ( ! is_admin() || ! $query->is_main_query() || Event::POST_TYPE !== $query->get( 'post_type' ) ) {
-			return;
-		}
-
-		if ( $column !== $query->get( 'orderby' ) ) {
-			return;
-		}
-
-		$order = strtoupper( $query->get( 'order', 'ASC' ) );
-		if ( ! in_array( $order, array( 'ASC', 'DESC' ), true ) ) {
-			$order = 'ASC';
-		}
-
-		foreach ( $filters as $hook => $callback ) {
-			add_filter( $hook, $callback );
-		}
-
-		$query->set( $order_key, $order );
-	}
-
-	/**
-	 * Join comments table for RSVP sorting (WordPress style).
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param string $join The JOIN clause of the query.
-	 * @return string Modified JOIN clause.
-	 */
-	public function rsvp_sorting_join_paged( string $join ): string {
-		global $wpdb;
-
-		$join .= $wpdb->prepare(
-			' LEFT JOIN %i AS rsvp_sort_comments'
-			. ' ON %i.%i = rsvp_sort_comments.comment_post_ID'
-			. ' AND rsvp_sort_comments.comment_type = %s'
-			. " AND rsvp_sort_comments.comment_approved = '1'",
-			$wpdb->comments,
-			$wpdb->posts,
-			'ID',
-			Rsvp::COMMENT_TYPE
-		);
-
-		return $join;
-	}
-
-	/**
-	 * Group by post ID for column sorting to prevent duplicate results.
-	 *
-	 * Shared callback used by both RSVP and venue sorting.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param string $groupby The GROUP BY clause of the query.
-	 * @return string Modified GROUP BY clause.
-	 */
-	public function sorting_groupby_post_id( string $groupby ): string {
-		global $wpdb;
-
-		if ( empty( $groupby ) ) {
-			$groupby = $wpdb->prepare( '%i.%i', $wpdb->posts, 'ID' );
-		}
-
-		return $groupby;
-	}
-
-	/**
-	 * Order by RSVP count for RSVP sorting (WordPress style).
-	 *
-	 * @since 1.0.0
-	 *
-	 * @return string Modified ORDER BY clause.
-	 */
-	public function rsvp_sorting_orderby(): string {
-		global $wp_query;
-
-		$order = $wp_query->get( 'rsvp_sort_order', 'ASC' );
-
-		// Remove the filters to prevent them from affecting other queries.
-		remove_filter( 'posts_join_paged', array( $this, 'rsvp_sorting_join_paged' ) );
-		remove_filter( 'posts_groupby', array( $this, 'sorting_groupby_post_id' ) );
-		remove_filter( 'posts_orderby', array( $this, 'rsvp_sorting_orderby' ) );
-
-		return "COUNT(rsvp_sort_comments.comment_ID) {$order}";
-	}
-
-	/**
-	 * Join term relationships and terms tables for venue sorting.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param string $join The JOIN clause of the query.
-	 * @return string Modified JOIN clause.
-	 */
-	public function venue_sorting_join_paged( string $join ): string {
-		global $wpdb;
-
-		$join .= $wpdb->prepare(
-			' LEFT JOIN %i AS venue_tr ON %i.%i = venue_tr.object_id',
-			$wpdb->term_relationships,
-			$wpdb->posts,
-			'ID'
-		);
-		$join .= $wpdb->prepare(
-			' LEFT JOIN %i AS venue_tt'
-			. ' ON venue_tr.term_taxonomy_id = venue_tt.term_taxonomy_id'
-			. ' AND venue_tt.taxonomy = %s',
-			$wpdb->term_taxonomy,
-			Venue::TAXONOMY
-		);
-		$join .= $wpdb->prepare(
-			' LEFT JOIN %i AS venue_terms ON venue_tt.term_id = venue_terms.term_id',
-			$wpdb->terms
-		);
-
-		return $join;
-	}
-
-	/**
-	 * Modify the ORDER BY clause for venue sorting.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @return string Modified ORDER BY clause.
-	 */
-	public function venue_sorting_orderby(): string {
-		global $wp_query;
-
-		$order = $wp_query->get( 'venue_sort_order', 'ASC' );
-
-		// Remove the filters to prevent them from affecting other queries.
-		remove_filter( 'posts_join_paged', array( $this, 'venue_sorting_join_paged' ) );
-		remove_filter( 'posts_groupby', array( $this, 'sorting_groupby_post_id' ) );
-		remove_filter( 'posts_orderby', array( $this, 'venue_sorting_orderby' ) );
-
-		// Sort by venue name, with NULL/empty values last.
-		return "CASE WHEN venue_terms.name IS NULL THEN 1 ELSE 0 END ASC, venue_terms.name {$order}";
 	}
 }
