+++ test/unit/php/includes/tests/core/classes/class-test-event-setup.php
@@ -85,34 +85,10 @@ class Test_Event_Setup extends Base {
 				'callback' => array( $instance, 'check_waiting_list' ),
 			),
 			array(
-				'type'     => 'action',
-				'name'     => sprintf( 'manage_%s_posts_custom_column', Event::POST_TYPE ),
-				'priority' => 10,
-				'callback' => array( $instance, 'custom_columns' ),
-			),
-			array(
-				'type'     => 'action',
-				'name'     => 'admin_menu',
-				'priority' => 10,
-				'callback' => array( $instance, 'modify_all_events_menu_link' ),
-			),
-			array(
-				'type'     => 'filter',
-				'name'     => 'submenu_file',
-				'priority' => 10,
-				'callback' => array( $instance, 'highlight_events_submenu' ),
-			),
-			array(
 				'type'     => 'filter',
 				'name'     => 'redirect_canonical',
 				'priority' => 10,
 				'callback' => array( $instance, 'disable_ics_canonical_redirect' ),
-			),
-			array(
-				'type'     => 'filter',
-				'name'     => sprintf( 'manage_%s_posts_columns', Event::POST_TYPE ),
-				'priority' => 10,
-				'callback' => array( $instance, 'set_custom_columns' ),
 			),
 			array(
 				'type'     => 'filter',
@@ -137,12 +113,6 @@ class Test_Event_Setup extends Base {
 				'name'     => 'display_post_states',
 				'priority' => 10,
 				'callback' => array( $instance, 'set_event_archive_labels' ),
-			),
-			array(
-				'type'     => 'filter',
-				'name'     => sprintf( 'manage_%s_posts_columns', Event::POST_TYPE ),
-				'priority' => 10,
-				'callback' => array( $instance, 'remove_comments_column' ),
 			),
 		);
 
@@ -454,70 +424,6 @@ class Test_Event_Setup extends Base {
 	}
 
 	/**
-	 * Coverage for remove_comments_column method.
-	 *
-	 * @covers ::remove_comments_column
-	 *
-	 * @return void
-	 */
-	public function test_remove_comments_column(): void {
-		$instance = Event_Setup::get_instance();
-
-		// Test with columns that include comments.
-		$columns_with_comments = array(
-			'cb'       => '<input type="checkbox" />',
-			'title'    => 'Title',
-			'comments' => 'Comments',
-			'date'     => 'Date',
-		);
-
-		$result = $instance->remove_comments_column( $columns_with_comments );
-
-		$expected = array(
-			'cb'    => '<input type="checkbox" />',
-			'title' => 'Title',
-			'date'  => 'Date',
-		);
-
-		$this->assertEquals(
-			$expected,
-			$result,
-			'Failed to assert that comments column is removed from the columns array.'
-		);
-
-		$this->assertArrayNotHasKey(
-			'comments',
-			$result,
-			'Failed to assert that comments key does not exist in result.'
-		);
-
-		// Test with columns that do not include comments.
-		$columns_without_comments = array(
-			'cb'    => '<input type="checkbox" />',
-			'title' => 'Title',
-			'date'  => 'Date',
-		);
-
-		$result = $instance->remove_comments_column( $columns_without_comments );
-
-		$this->assertEquals(
-			$columns_without_comments,
-			$result,
-			'Failed to assert that columns without comments remain unchanged.'
-		);
-
-		// Test with empty array.
-		$empty_columns = array();
-		$result        = $instance->remove_comments_column( $empty_columns );
-
-		$this->assertEquals(
-			$empty_columns,
-			$result,
-			'Failed to assert that empty array remains unchanged.'
-		);
-	}
-
-	/**
 	 * Coverage for check_waiting_list method.
 	 *
 	 * @covers ::check_waiting_list
@@ -589,230 +495,6 @@ class Test_Event_Setup extends Base {
 		// phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching
 		$count = $wpdb->get_var( $wpdb->prepare( 'SELECT COUNT(*) FROM %i WHERE post_id = %d', $table, $post_id ) );
 		$this->assertEquals( 0, $count, 'Event record should be deleted from custom table.' );
-	}
-
-	/**
-	 * Coverage for custom_columns method with datetime column.
-	 *
-	 * @covers ::custom_columns
-	 *
-	 * @return void
-	 */
-	public function test_custom_columns_datetime(): void {
-		$instance = Event_Setup::get_instance();
-		$post_id  = $this->mock->post(
-			array( 'post_type' => Event::POST_TYPE )
-		)->get()->ID;
-
-		$event = new Event( $post_id );
-		$event->save_datetimes(
-			array(
-				'datetime_start' => '2025-06-15 10:00:00',
-				'datetime_end'   => '2025-06-15 14:00:00',
-				'timezone'       => 'America/New_York',
-			)
-		);
-
-		ob_start();
-		$instance->custom_columns( 'datetime', $post_id );
-		$output = ob_get_clean();
-
-		$this->assertNotEmpty( $output, 'Datetime column should produce output.' );
-		$this->assertStringContainsString( 'June', $output, 'Datetime output should contain month name.' );
-	}
-
-	/**
-	 * Coverage for custom_columns method with venue column.
-	 *
-	 * @covers ::custom_columns
-	 *
-	 * @return void
-	 */
-	public function test_custom_columns_venue_with_name(): void {
-		$instance = Event_Setup::get_instance();
-		$post_id  = $this->mock->post(
-			array( 'post_type' => Event::POST_TYPE )
-		)->get()->ID;
-
-		// Create a venue term and associate it with the event.
-		$venue_name = 'Test Venue';
-		$term       = wp_insert_term( $venue_name, \GatherPress\Core\Venue::TAXONOMY );
-		wp_set_post_terms( $post_id, array( $term['term_id'] ), \GatherPress\Core\Venue::TAXONOMY );
-
-		ob_start();
-		$instance->custom_columns( 'venue', $post_id );
-		$output = ob_get_clean();
-
-		$this->assertStringContainsString( $venue_name, $output, 'Venue column should contain venue name.' );
-	}
-
-	/**
-	 * Coverage for custom_columns method with venue column and no venue.
-	 *
-	 * @covers ::custom_columns
-	 *
-	 * @return void
-	 */
-	public function test_custom_columns_venue_no_venue(): void {
-		$instance = Event_Setup::get_instance();
-		$post_id  = $this->mock->post(
-			array( 'post_type' => Event::POST_TYPE )
-		)->get()->ID;
-
-		ob_start();
-		$instance->custom_columns( 'venue', $post_id );
-		$output = ob_get_clean();
-
-		$this->assertStringContainsString( '—', $output, 'Venue column should show em dash when no venue.' );
-	}
-
-	/**
-	 * Coverage for custom_columns method with venue column and online event.
-	 *
-	 * @covers ::custom_columns
-	 *
-	 * @return void
-	 */
-	public function test_custom_columns_venue_online_event(): void {
-		$instance = Event_Setup::get_instance();
-		$post_id  = $this->mock->post(
-			array( 'post_type' => Event::POST_TYPE )
-		)->get()->ID;
-
-		// Set online event link.
-		update_post_meta( $post_id, 'gatherpress_online_event_link', 'https://example.com/meeting' );
-
-		ob_start();
-		$instance->custom_columns( 'venue', $post_id );
-		$output = ob_get_clean();
-
-		// The method should execute without error for online events.
-		$this->assertNotEmpty( $output, 'Online event column should produce output.' );
-	}
-
-	/**
-	 * Coverage for custom_columns method with rsvps column (no RSVPs).
-	 *
-	 * @covers ::custom_columns
-	 *
-	 * @return void
-	 */
-	public function test_custom_columns_rsvps_no_rsvps(): void {
-		$instance = Event_Setup::get_instance();
-		$post_id  = $this->mock->post(
-			array( 'post_type' => Event::POST_TYPE )
-		)->get()->ID;
-
-		ob_start();
-		$instance->custom_columns( 'rsvps', $post_id );
-		$output = ob_get_clean();
-
-		$this->assertStringContainsString( '—', $output, 'RSVPs column should show em dash when no RSVPs.' );
-	}
-
-	/**
-	 * Coverage for custom_columns method with rsvps column (with approved RSVPs).
-	 *
-	 * @covers ::custom_columns
-	 *
-	 * @return void
-	 */
-	public function test_custom_columns_rsvps_with_approved(): void {
-		$instance = Event_Setup::get_instance();
-		$post_id  = $this->mock->post(
-			array( 'post_type' => Event::POST_TYPE )
-		)->get()->ID;
-
-		// Create approved RSVP.
-		$this->factory->comment->create(
-			array(
-				'comment_post_ID'  => $post_id,
-				'comment_type'     => \GatherPress\Core\Rsvp::COMMENT_TYPE,
-				'comment_approved' => 1,
-			)
-		);
-
-		ob_start();
-		$instance->custom_columns( 'rsvps', $post_id );
-		$output = ob_get_clean();
-
-		$this->assertStringContainsString( 'gatherpress-rsvp-approved', $output, 'Should show approved RSVP count.' );
-		$this->assertStringContainsString( '>1<', $output, 'Should show count of 1.' );
-	}
-
-	/**
-	 * Coverage for custom_columns method with rsvps column (with unapproved RSVPs).
-	 *
-	 * @covers ::custom_columns
-	 *
-	 * @return void
-	 */
-	public function test_custom_columns_rsvps_with_unapproved(): void {
-		$instance = Event_Setup::get_instance();
-		$post_id  = $this->mock->post(
-			array( 'post_type' => Event::POST_TYPE )
-		)->get()->ID;
-
-		// Create approved and unapproved RSVPs.
-		$this->factory->comment->create(
-			array(
-				'comment_post_ID'  => $post_id,
-				'comment_type'     => \GatherPress\Core\Rsvp::COMMENT_TYPE,
-				'comment_approved' => 1,
-			)
-		);
-		$this->factory->comment->create(
-			array(
-				'comment_post_ID'  => $post_id,
-				'comment_type'     => \GatherPress\Core\Rsvp::COMMENT_TYPE,
-				'comment_approved' => 0,
-			)
-		);
-
-		ob_start();
-		$instance->custom_columns( 'rsvps', $post_id );
-		$output = ob_get_clean();
-
-		$this->assertStringContainsString(
-			'gatherpress-rsvp-pending',
-			$output,
-			'Should show unapproved RSVP indicator.'
-		);
-		$this->assertStringContainsString( 'Unapproved RSVPs', $output, 'Should contain title for unapproved.' );
-	}
-
-	/**
-	 * Coverage for set_custom_columns method.
-	 *
-	 * @covers ::set_custom_columns
-	 *
-	 * @return void
-	 */
-	public function test_set_custom_columns(): void {
-		$instance = Event_Setup::get_instance();
-
-		$default_columns = array(
-			'cb'     => '<input type="checkbox" />',
-			'title'  => 'Title',
-			'author' => 'Author',
-			'date'   => 'Date',
-		);
-
-		$result = $instance->set_custom_columns( $default_columns );
-
-		// Should not contain author column.
-		$this->assertArrayNotHasKey( 'author', $result, 'Author column should be removed.' );
-
-		// Should contain custom columns.
-		$this->assertArrayHasKey( 'datetime', $result, 'Should have datetime column.' );
-		$this->assertArrayHasKey( 'venue', $result, 'Should have venue column.' );
-		$this->assertArrayHasKey( 'rsvps', $result, 'Should have rsvps column.' );
-
-		// Verify order (custom columns should be inserted after title).
-		$keys = array_keys( $result );
-		$this->assertEquals( 'cb', $keys[0], 'First column should be cb.' );
-		$this->assertEquals( 'title', $keys[1], 'Second column should be title.' );
-		$this->assertEquals( 'datetime', $keys[2], 'Third column should be datetime.' );
 	}
 
 	/**
@@ -1443,139 +1125,6 @@ class Test_Event_Setup extends Base {
 		$this->assertStringContainsString( 'BEGIN:VEVENT', $output );
 		$this->assertStringContainsString( 'END:VEVENT', $output );
 		$this->assertStringContainsString( 'END:VCALENDAR', $output );
-	}
-
-	/**
-	 * Tests custom_columns with online event.
-	 *
-	 * Covers: is_online_event icon display.
-	 *
-	 * @covers ::custom_columns
-	 * @return void
-	 */
-	public function test_custom_columns_venue_with_online_event(): void {
-		$post_id = $this->mock->post( array( 'post_type' => Event::POST_TYPE ) )->get()->ID;
-
-		// Create 'online-event' venue term.
-		$term = wp_insert_term(
-			'Online Event',
-			'_gatherpress_venue',
-			array( 'slug' => 'online-event' )
-		);
-
-		// Assign the term to the event.
-		wp_set_object_terms( $post_id, $term['term_id'], '_gatherpress_venue' );
-
-		$instance = Event_Setup::get_instance();
-
-		ob_start();
-		$instance->custom_columns( 'venue', $post_id );
-		$output = ob_get_clean();
-
-		$this->assertStringContainsString( 'dashicons-video-alt3', $output, 'Should display online event icon' );
-	}
-
-	/**
-	 * Coverage for modify_all_events_menu_link method.
-	 *
-	 * @covers ::modify_all_events_menu_link
-	 *
-	 * @return void
-	 */
-	public function test_modify_all_events_menu_link(): void {
-		global $submenu;
-
-		$instance  = Event_Setup::get_instance();
-		$menu_slug = sprintf( 'edit.php?post_type=%s', Event::POST_TYPE );
-
-		// Set up a mock submenu structure.
-		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$submenu[ $menu_slug ] = array(
-			5  => array( 'View Events', 'edit_posts', $menu_slug ),
-			10 => array( 'Add New', 'edit_posts', 'post-new.php?post_type=' . Event::POST_TYPE ),
-		);
-
-		$instance->modify_all_events_menu_link();
-
-		$this->assertStringContainsString(
-			'gatherpress_event_query=upcoming',
-			$submenu[ $menu_slug ][5][2],
-			'Upcoming Events menu link should include upcoming parameter.'
-		);
-
-		// Other submenu items should be unchanged.
-		$this->assertStringNotContainsString(
-			'gatherpress_event_query',
-			$submenu[ $menu_slug ][10][2],
-			'Add New menu link should not be modified.'
-		);
-
-		// Clean up.
-		unset( $submenu[ $menu_slug ] );
-	}
-
-	/**
-	 * Coverage for modify_all_events_menu_link method with empty submenu.
-	 *
-	 * @covers ::modify_all_events_menu_link
-	 *
-	 * @return void
-	 */
-	public function test_modify_all_events_menu_link_empty_submenu(): void {
-		global $submenu;
-
-		$instance  = Event_Setup::get_instance();
-		$menu_slug = sprintf( 'edit.php?post_type=%s', Event::POST_TYPE );
-
-		// Ensure the submenu doesn't exist.
-		unset( $submenu[ $menu_slug ] );
-
-		// Should return early without error.
-		$instance->modify_all_events_menu_link();
-
-		$this->assertFalse(
-			isset( $submenu[ $menu_slug ] ),
-			'Should not create submenu when none exists.'
-		);
-	}
-
-	/**
-	 * Coverage for highlight_events_submenu when on events page.
-	 *
-	 * @covers ::highlight_events_submenu
-	 *
-	 * @return void
-	 */
-	public function test_highlight_events_submenu_on_events_page(): void {
-		$instance  = Event_Setup::get_instance();
-		$menu_slug = sprintf( 'edit.php?post_type=%s', Event::POST_TYPE );
-
-		$result = $instance->highlight_events_submenu( $menu_slug );
-
-		$this->assertStringContainsString(
-			'gatherpress_event_query=upcoming',
-			$result,
-			'Should return modified slug with upcoming parameter.'
-		);
-	}
-
-	/**
-	 * Coverage for highlight_events_submenu on a different page.
-	 *
-	 * @covers ::highlight_events_submenu
-	 *
-	 * @return void
-	 */
-	public function test_highlight_events_submenu_other_page(): void {
-		$instance = Event_Setup::get_instance();
-
-		$result = $instance->highlight_events_submenu( 'edit.php?post_type=post' );
-
-		$this->assertSame(
-			'edit.php?post_type=post',
-			$result,
-			'Should not modify submenu file for other post types.'
-		);
 	}
 
 	/**
